model/requirements.txt

# Core dependencies
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
pyjwt==2.6.0
redis==4.5.1
psutil==5.9.4

# ML dependencies
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2

# Optional ML dependencies
# Uncomment as needed for your specific model
# tensorflow==2.11.0
# torch==2.0.0
# xgboost==1.7.4
# lightgbm==3.3.5

# Monitoring
prometheus-flask-exporter==0.22.3
